requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
django-jalali==6.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
ippanel==2.0.7
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
